Update update_server.py
Iterating through the lines we are using a mod calculation to determine if the number of items in hash_list is divisible by HASH_LEN, which by default == 1000. This approach is good for memory allocation as well as network bandwidth; however, it will neglect whatever items remain after the last time hash_list IS divisible by 1000. For example: If we have 2,500 items, the if check using mod % will succeed 2 times putting the first 2,000 items from the file into safelist. The remaining 500 items will be placed into hash_list, but never make it to a client._connection.put call to put them into safelist. This additional if statement solves and handles the remaining 500 items such that all hashes found in source text are put into safelist appropriately.